+++ src/main/java/kitchenpos/application/MenuGroupService.java
@@ -7,7 +7,7 @@ import org.springframework.transaction.annotation.Transactional;
 
 import kitchenpos.dao.MenuGroupDao;
 import kitchenpos.domain.MenuGroup;
-import kitchenpos.domain.MenuGroupRequest;
+import kitchenpos.dto.MenuGroupRequest;
 
 @Service
 @Transactional(readOnly = true)
