+++ src/main/java/kitchenpos/dto/MenuGroupRequest.java
@@ -1,4 +1,6 @@
-package kitchenpos.domain;
+package kitchenpos.dto;
+
+import kitchenpos.domain.MenuGroup;
 
 public class MenuGroupRequest {
     private String name;
