+++ src/main/java/kitchenpos/dto/MenuRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.dto;
 
 import java.math.BigDecimal;
 import java.util.List;
