+++ src/main/java/kitchenpos/dto/ProductResponse.java
@@ -1,7 +1,9 @@
-package kitchenpos.domain;
+package kitchenpos.dto;
 
 import java.math.BigDecimal;
 import java.util.Objects;
+
+import kitchenpos.domain.Product;
 
 public class ProductResponse {
     private Long id;
