+++ src/test/java/kitchenpos/application/MenuServiceTest.java
@@ -23,10 +23,10 @@ import kitchenpos.dao.ProductDao;
 import kitchenpos.domain.Menu;
 import kitchenpos.domain.MenuGroup;
 import kitchenpos.domain.MenuProduct;
-import kitchenpos.domain.MenuProductRequest;
-import kitchenpos.domain.MenuProductResponse;
-import kitchenpos.domain.MenuRequest;
-import kitchenpos.domain.MenuResponse;
+import kitchenpos.dto.MenuProductRequest;
+import kitchenpos.dto.MenuProductResponse;
+import kitchenpos.dto.MenuRequest;
+import kitchenpos.dto.MenuResponse;
 import kitchenpos.domain.Product;
 import kitchenpos.exception.KitchenposException;
 import kitchenpos.exception.KitchenposNotFoundException;
