+++ src/test/java/kitchenpos/application/ProductServiceTest.java
@@ -16,8 +16,8 @@ import org.mockito.junit.jupiter.MockitoExtension;
 
 import kitchenpos.dao.ProductDao;
 import kitchenpos.domain.Product;
-import kitchenpos.domain.ProductRequest;
-import kitchenpos.domain.ProductResponse;
+import kitchenpos.dto.ProductRequest;
+import kitchenpos.dto.ProductResponse;
 import kitchenpos.exception.KitchenposException;
 
 @ExtendWith(MockitoExtension.class)
